Add base ViewModel class
--In order to share common methods from TweetListVM - move it in BaseViewModel

[Ticket: baseViewModel]
