+++ app/src/main/java/com/sa/betvictor/ui/TweetListViewModel.kt
@@ -1,11 +1,8 @@
 package com.sa.betvictor.ui
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import com.sa.betvictor.R
-import com.sa.betvictor.common.ActionLiveData
+import com.sa.betvictor.app.base.BaseViewModel
 import com.sa.betvictor.common.NetworkStateMonitor
 import com.sa.betvictor.common.Timer
 import com.sa.betvictor.common.Timer.OnScheduledTimerExpiredListener
@@ -14,45 +11,31 @@ import com.sa.betvictor.domain.TweetRepository
 import com.sa.betvictor.ui.TweetListViewModel.FetchState.ACTIVE
 import com.sa.betvictor.ui.TweetListViewModel.FetchState.INACTIVE
 import com.sa.betvictor.ui.TweetQueryValidator.TweetQueryValidationListener
-import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.launch
-import okhttp3.internal.http2.StreamResetException
-import java.net.UnknownHostException
 
 class TweetListViewModel(
     private val repository: TweetRepository,
     private val validator: TweetQueryValidator,
-    private val networkMonitor: NetworkStateMonitor,
-    private val timer: Timer
-) : ViewModel(), TweetQueryValidationListener, OnScheduledTimerExpiredListener {
+    private val timer: Timer,
+    networkMonitor: NetworkStateMonitor
+) : BaseViewModel(networkMonitor), TweetQueryValidationListener, OnScheduledTimerExpiredListener {
 
-    private val TAG = TweetListViewModel::class.java.simpleName
-    private val viewModelJob = Job()
-    private val uiScope = CoroutineScope(Dispatchers.Main + viewModelJob)
+    override val tag: String = TweetListViewModel::class.java.simpleName
     private var fetchTweetsJob: Job? = null
 
     private val _fetchState = MutableLiveData(INACTIVE)
     private val _tweetData = MutableLiveData<List<Tweet>>()
-    private val _progress = MutableLiveData<Boolean>()
     private val _onInvalidQuery = MutableLiveData<Int>()
-    private val _error = ActionLiveData<Int>()
 
     val fetchState: LiveData<FetchState> = _fetchState
     val tweetData: LiveData<List<Tweet>> = _tweetData
-    val progress: LiveData<Boolean> = _progress
     val onInvalidQuery: LiveData<Int> = _onInvalidQuery
-    val error: LiveData<Int> = _error
 
     init {
         validator.listener = this
         timer.setOnPeriodFinishedListener(this)
     }
-
-    fun registerNetworkCallback() = networkMonitor.registerNetworkCallback()
 
     fun getTweets() {
         launchDataLoad {
@@ -81,15 +64,8 @@ class TweetListViewModel(
         }
     }
 
-    private fun onFetchTweetsFails(t: Throwable) {
-        _progress.value = false
-        _fetchState.value = INACTIVE
-        timer.stop()
-    }
-
-    private fun scheduleTimer() {
-        val period = if (TWEET_LIFESPAN > ONE_SECOND) ONE_SECOND else TWEET_LIFESPAN
-        timer.schedule(uiScope, period)
+    fun onNetworkAvailable() {
+        scheduleTimer()
     }
 
     override fun onInvalidQuery(errorResId: Int) {
@@ -102,41 +78,15 @@ class TweetListViewModel(
         if (deletedTweets == 0 && fetchTweetsJob?.isActive != true) timer.stop()
     }
 
-    private fun launchDataLoad(
-        doOnError: (Throwable) -> Unit = { },
-        doOnComplete: () -> Unit = { },
-        block: suspend () -> Unit
-    ): Job = uiScope.launch {
-        try {
-            block()
-        } catch (t: Throwable) {
-            doOnError(t)
-            onError(t)
-        } finally {
-            doOnComplete()
-        }
+    private fun onFetchTweetsFails(t: Throwable) {
+        _progress.value = false
+        _fetchState.value = INACTIVE
+        timer.stop()
     }
 
-    private fun onError(throwable: Throwable) {
-        when (throwable) {
-            is CancellationException -> Unit
-
-            //Todo - error handling needed - Http, Twitter API errors and edge cases
-            is StreamResetException -> _error.value = R.string.error_stream_canceled
-            is UnknownHostException -> _error.value = R.string.error_no_network_connection
-            else -> _error.value = R.string.error_generic
-        }
-        Log.e(TAG, throwable.message, throwable)
-    }
-
-    override fun onCleared() {
-        super.onCleared()
-        networkMonitor.unregisterNetworkCallback()
-        viewModelJob.cancel()
-    }
-
-    fun onNetworkAvailable() {
-        scheduleTimer()
+    private fun scheduleTimer() {
+        val period = if (TWEET_LIFESPAN > ONE_SECOND) ONE_SECOND else TWEET_LIFESPAN
+        timer.schedule(uiScope, period)
     }
 
     enum class FetchState { ACTIVE, INACTIVE }
